+++ R/build-site.R
@@ -16,6 +16,8 @@ build_site <- function(path = ".", quiet = TRUE, ...) {
 
   pkg <- as.package(path)
 
+  check_dots_used()
+
   withr::with_temp_libpaths(action = "prefix", code = {
     install(pkg = pkg$path, upgrade = "never", reload = FALSE, quiet = quiet)
     if (isTRUE(quiet)) {