+++ R/pkgbuild.R
@@ -7,6 +7,9 @@
 build <- function(pkg = ".", path = NULL, binary = FALSE, vignettes = TRUE,
                   manual = FALSE, args = NULL, quiet = FALSE, ...) {
   save_all()
+
+  check_dots_used()
+
   pkgbuild::build(
     path = pkg, dest_path = path, binary = binary,
     vignettes = vignettes, manual = manual, args = args, quiet = quiet, ...